get.summary.table {QoRTs}	R Documentation

Get summary data table

Description

Retrieves and compiles a summary data table.

Usage

  get.summary.table(res, outfile, debugMode)

Arguments

res
A QoRT_QC_Results object, created by read.qc.results.data.

outfile
Optional. A file name where the table should be written.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

Details

Returns summary data in table form.

See Also

read.qc.results.data

[Package QoRTs version 0.1.7 Index]