makePlot.all.std {QoRTs}	R Documentation

Generating all default plots

Description

Saves MANY compiled QC plots for the given dataset.

Usage

  makeMultiPlot.all(res, outfile.dir = "./", 
                     plotter.params = list(), 
                     plot.device.name = "png", 
                     plotting.device.params = list(), 
                     debugMode,
                     rasterize.large.plots,
                     raster.height = 1000, 
                     raster.width = 1000,
                     exclude.autosomes.chrom.rate.plot = TRUE,
                     chromosome.name.style = "UCSC",
                     fig.res = 150, fig.base.height.inches = 7,
                     ...)

Arguments

res
A QoRT_QC_Results object, created by read.qc.results.data.

outfile.dir
A file prefix, used for all output files. Usually the directory to which you want all files to be written.

plotter.params
Additional (advanced) parameters used in creation of the Plotter objects. See build.plotter.

plot.device.name
The method to use to save plots. Can be one of:

"png" for standard png compression,

"CairoPNG" for png compression using package Cairo. Note that this requires the package Cairo.

plotting.device.params
A named list of parameters to be passed to the graphics device. For example:

"width = 2000"

Reasonable values for height, width, and pointsize will be chosen by default.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

rasterize.large.plots
Logical. If TRUE, then if the currently selected plotting device is a vector device (or the "curr" device), then certain large plots will have their plotting areas rasterized. The axis labels, titles, and text will all remain vectorized, only the plotting areas will be flattened. Note that this requires the png package.

By default, this parameter will be set to TRUE when a vector device is selected.

raster.height
Numeric. If rasterize.plotting.area is TRUE, then this is the height of the plotting area raster image, in pixels.

raster.width
Numeric. If rasterize.plotting.area is TRUE, then this is the width of the plotting area raster image, in pixels. Double-pane plots will be twice this width.

exclude.autosomes.chrom.rate.plot
A logical value indicating whether to exclude autosomes from the plot. See makePlot.chrom.type.rates

chromosome.name.style
A string value indicating the style of the chromosome names, and also how to split up the categories. See makePlot.chrom.type.rates

fig.res
Numeric value. The number of pixels per "inch" (for raster devices only). For some plotting devices the figure height will be in pixels not inches, and will equal this value multiplied by the fig.base.height.inches value.

fig.base.height.inches
Numeric value. The base height, in inches, of each sub-figure in the plot. This will be equal to the height for vector devices, or used to calculate the height in pixels using the fig.res value (see above).

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

Saves MANY compiled QC plots for the given dataset, calling all standard variants of makeMultiPlot.

See Also

read.qc.results.data, build.plotter, makeMultiPlot

[Package QoRTs version 0.1.7 Index]