makeMultiPlot.highlightSample.all {QoRTs}	R Documentation

Generating sample highlight multi-plots

Description

Generates multiple sample highlight summary plots, one for every sample.

Usage

  makeMultiPlot.highlightSample.all(res, outfile.dir = "./",
                plotter.params = list(), 
                plot.device.name = "png", 
                plotting.device.params = list(), 
                verbose = TRUE,
                debugMode , 
                rasterize.large.plots, 
                rasterize.medium.plots,
                raster.height = 1050, 
                raster.width = 1050,
                exclude.autosomes.chrom.rate.plot = TRUE,
                chromosome.name.style = "UCSC",
                fig.res = 150, fig.base.height.inches = 7,
                insertSize.plot.xlim,
                sequencing.type = c("RNA","Exome","Genome"),
                maxColumns,
                maxRows,
                plotList,
                labelPlots=TRUE,
                ...)

makeMultiPlot.highlightSample.colorByLane.all(res, 
                outfile.dir = "./",
                plotter.params = list(), 
                plot.device.name = "png", 
                plotting.device.params = list(), 
                verbose = TRUE, 
                debugMode ,
                rasterize.large.plots, 
                rasterize.medium.plots,
                raster.height = 1050, 
                raster.width = 1050,
                exclude.autosomes.chrom.rate.plot = TRUE,
                chromosome.name.style = "UCSC",
                fig.res = 150, fig.base.height.inches = 7,
                insertSize.plot.xlim,
                sequencing.type = c("RNA","Exome","Genome"),
                maxColumns,
                maxRows,
                plotList,
                labelPlots=TRUE,
                ...)

Arguments

res
A QoRT_QC_Results object, created by read.qc.results.data.

outfile.dir
A file prefix, used for all output files. Usually the directory to which you want all files to be written.

plotter.params
Additional parameters (advanced) used in creation of the Plotter objects. See build.plotter.

plot.device.name
The method to use to save plots. Can be one of:

"png" for standard png compression,

"CairoPNG" for png compression using package Cairo. Note that this requires the package Cairo.

plotting.device.params
A named list of parameters to be passed to the graphics device. For example:

"width = 2000"

Reasonable values for height, width, and pointsize will be chosen by default.

verbose
Logical. If TRUE, more info will be printed to the console.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

rasterize.large.plots
Logical. If TRUE, then if the currently selected plotting device is a vector device (or the "curr" device), then certain large plots will have their plotting areas rasterized. The axis labels, titles, and text will all remain vectorized, only the plotting areas will be flattened. Note that this requires the png package.

By default, this parameter will be set to TRUE when a vector device is selected.

rasterize.medium.plots
Logical. as rasterize.large.plots, but applies to moderately-large plots. By default, this parameter will be set to TRUE for pdf/CairoPDF output only.

raster.height
Numeric. If rasterize.plotting.area is TRUE, then this is the height of the plotting area raster image, in pixels.

raster.width
Numeric. If rasterize.plotting.area is TRUE, then this is the width of the plotting area raster image, in pixels. Double-pane plots will be twice this width.

exclude.autosomes.chrom.rate.plot
A logical value indicating whether to exclude autosomes from the plot. See makePlot.chrom.type.rates

chromosome.name.style
A string value indicating the style of the chromosome names, and also how to split up the categories. See makePlot.chrom.type.rates

fig.res
Numeric value. The number of pixels per "inch" (for raster devices only). For some plotting devices the figure height will be in pixels not inches, and will equal this value multiplied by the fig.base.height.inches value.

fig.base.height.inches
Numeric value. The base height, in inches, of each sub-figure in the plot. This will be equal to the height for vector devices, or used to calculate the height in pixels using the fig.res value (see above).

insertSize.plot.xlim
A numeric vector of length 2. The x-axis limits for the insert size plot. By default QoRTs will attempt to find reasonable values for this, but there are always situations where the default behavior is not ideal. Using this parameter you can set it explicitly.

sequencing.type
The type of sequencing data being analyzed. This only changes the default plot set, which can be overriden with the plotList parameter.

maxColumns
If set, QoRTs will attempt to create a multiplot that has (at most) maxColumns columns. Extra rows will be added to fit all the plots, as needed.

maxRows
If set, QoRTs will attempt to create a multiplot that has (at most) maxRows rows. Extra columns will be added to fit all the plots as needed. To set the number of rows and columns manually, you can set both maxColumns and maxRows.

plotList
The list of desired plots.

labelPlots
Logical. If TRUE, then label each plot with a letter.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

Generates a sample highlight plot for each sample in the dataset.

See Also

build.plotter, makeMultiPlot

Examples

  ## Not run: 
##D   data(res,package="QoRTsExampleData");
##D     makeMultiPlot.highlightSample.all(res);
##D     makeMultiPlot.highlightSample.colorByLane.all(res);
##D   
## End(Not run)

[Package QoRTs version 1.2.19 Index]