makePlot.NVC.clip.matchByClipPosition {QoRTs}	R Documentation

Plot clipped nucleotide rates, aligned by the distance from the point of clipping.

Description

WARNING: THESE FUNCTIONS ARE BETA, AND ARE NOT FULLY TESTED OR READY FOR GENERAL USE.

Usage

makePlot.NVC.lead.clip.matchByClipPosition(plotter, 
            clip.amt,  r2.buffer, 
            label.majority.bases = TRUE, 
            label.majority.bases.threshold = 0.5, 
            label.majority.bases.cex = 1, 
            rasterize.plotting.area = FALSE, 
            raster.height = 1000, 
            raster.width = 1000,
            show.base.legend = TRUE, 
            load.results = TRUE, 
            debugMode, singleEndMode, ...)

makePlot.NVC.tail.clip.matchByClipPosition(plotter, 
            clip.amt, r2.buffer,
            label.majority.bases = TRUE, 
            label.majority.bases.threshold = 0.5, 
            label.majority.bases.cex = 1, 
            rasterize.plotting.area = FALSE, 
            raster.height = 1000, 
            raster.width = 1000,
            show.base.legend = TRUE, 
            debugMode, singleEndMode, ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

clip.amt
UNDOCUMENTED

r2.buffer
UNDOCUMENTED

label.majority.bases
UNDOCUMENTED

label.majority.bases.threshold
UNDOCUMENTED

label.majority.bases.cex
UNDOCUMENTED

show.base.legend
UNDOCUMENTED

rasterize.plotting.area
Logical. If TRUE, the plotting area will be written to a temp png file then drawn to the current device as a raster image. This option is generally preferred for vector devices, because NVC plots can be very large when drawn in vector format. Note: this requires the png package!

raster.height
Numeric. If rasterize.plotting.area is TRUE, then this is the height of the plotting area raster image, in pixels.

raster.width
Numeric. If rasterize.plotting.area is TRUE, then this is the width of the plotting area raster image, in pixels.

load.results
UNDOCUMENTED

debugMode
UNDOCUMENTED

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

This function is currently BETA, and is not intended for general use. Documentation and testing is still pending.

See Also

build.plotter, makePlot.NVC

[Package QoRTs version 0.1.6 Index]