makePlot.biotype.rates {QoRTs}	R Documentation

Plot Biotype rates

Description

Plots counts for each gene biotype. This plot is only useful and informative when QoRTs is run on a GTF file that contains "gene_biotype" tags.

Usage

  makePlot.biotype.rates(plotter, 
              plot.rates = TRUE,
              count.type = c("all","unambigOnly"),
              log.y = TRUE,
              return.table = FALSE, 
              debugMode = DEFAULTDEBUGMODE,  
              singleEndMode = plotter$res@singleEnd, ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

plot.rates
A logical value. If TRUE, then reads/read-pairs per million should be plotted, rather than raw counts. Default is TRUE.

count.type
A logical value. If TRUE, then both ambiguous and unambiguous reads will be counted. Otherwise only unambiguous reads will be counted.

log.y
A logical value indicating that the y-axis should be log-scaled.

return.table
Logical. If TRUE, then return a data table.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

x-axis: Gene "BioTypes", from the annotation GTF.

y-axis: Total read counts across all genes in the BioType, measured either in reads or reads-per-million.

See Also

build.plotter

[Package QoRTs version 1.0.0 Index]