makePlot.cigarOp.byCycle {QoRTs}	R Documentation

Plot Cigar Operator Rate

Description

Plots the rate at which the given CIGAR operator occurs, by read cycle.

Usage

  makePlot.cigarOp.byCycle(plotter, op, 
                       r2.buffer = NULL, 
                       rate.per.million = TRUE, 
                       debugMode, singleEndMode,
                       ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

op
A character string naming which cigar op to plot. Must be one of the following:

"SoftClip" Soft Clip (Cigar Op "S")

"HardClip" Hard Clip (Cigar Op "H")

"Del" Deletion from reference (Cigar Op "D")

"Ins" Insertion from reference (Cigar Op "I")

"Pad" Padding (Cigar Op "P")

"Splice" Splice Junction (Cigar Op "N")

"Aln" Aligned to reference (Cigar Op "M")

Note that makePlot.cigarOp.byCycle(plotter,"SoftClip") gives identical results as makePlot.clipping, although the default value for the rate.per.million argument is different.

r2.buffer
Buffer space to place between the plotting of read 1 and read 2. By default this will choose a reasonable value.

rate.per.million
A logical value indicating whether or not to scale the y axis to rate-per-million-reads, rather than rate-per-read. Some people may find the results more readable this way, even though the plots themselves will appear the same.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

x-axis: The read cycle (ie. the base-pair position in the read).

y-axis: The rate at which the bases at the given read-cycle is clipped off.

See Also

build.plotter

[Package QoRTs version 0.3.3 Index]