makePlot.clipping {QoRTs}	R Documentation

Plot Alignment Clipping

Description

Plots the rate at which the aligner soft-clips off portions of aligned reads.

Usage

  makePlot.clipping(plotter, rate.per.million = FALSE, 
                r2.buffer = NULL, 
                debugMode, singleEndMode,
                ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

rate.per.million
A logical value indicating whether or not to scale the y axis to rate-per-million-reads, rather than rate-per-read. Some people may find the results more readable this way, even though the plots themselves will appear the same.

r2.buffer
Buffer space to place between the plotting of read 1 and read 2. By default this will choose a reasonable value.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

x-axis: The read cycle (ie. the base-pair position in the read).

y-axis: The rate at which the bases at the given read-cycle is clipped off.

See Also

build.plotter

[Package QoRTs version 0.3.26 Index]