makePlot.dropped.rates {QoRTs}	R Documentation

Read Drop Plot

Description

Plots the rates at which reads are dropped from analysis for various causes.

Usage

makePlot.dropped.rates(plotter, dropAlwaysZeroRows = FALSE, 
                       debugMode, singleEndMode, ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

dropAlwaysZeroRows
Logical. If TRUE, drop-reasons that never occur in the dataset will not be plotted. This often cleans up the plot somewhat, since in many production pipelines reads that fail many of the filtering steps may have already been filtered out.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

For each bam file, this function plots the rates and reasons for reads being dropped from QC analysis.

Note that in the example dataset reads were never dropped. This is a consequence of the pre-processing steps in the example pipeline.

See Also

build.plotter

[Package QoRTs version 0.1.6 Index]