makePlot.gc {QoRTs}	R Documentation

Plot GC content

Description

Plots GC content.

Usage

  makePlot.gc(plotter, plot.medians = NULL, plot.means = TRUE, 
              plotRate = FALSE, byPair = FALSE,
              debugMode, singleEndMode, ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

plot.medians
A logical value indicating whether or not to plot the median average GC content for each bam file at the bottom of the plot. Overrides plot.means.

plot.means
A logical value indicating whether or not to plot the mean average GC content for each bam file at the bottom of the plot.

plotRate
A logical value indicating whether or not the X-axis should be the raw number of nucleotides that are G/C, vs the rate G/C.

byPair
Logical. If FALSE, GC content rates will be calculated for all reads individually. If TRUE, GC content rates will be calculated for all read-pairs. Note that when the insert size is small, the paired plot can sometimes appear jagged, as completely-overlapped read-pairs will always have an even number of G/C nucleotides.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

x-axis: Percent of the read-pairs that is made up of G's or C's.

y-axis: Rate at which the given percent occurs.

See Also

build.plotter

[Package QoRTs version 0.2.13 Index]