makePlot.gene.assignment.rates {QoRTs}	R Documentation

Gene assignment rate plot

Description

Plots the rate at which reads are assigned into various categories.

Usage

makePlot.gene.assignment.rates(plotter,  debugMode, singleEndMode, ...)
                         

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

For each bam-file, this function plots the rate (y-axis) for which the bam-file's read-pairs are assigned to the given categories.

The categories are:

Unique Gene: The read-pair overlaps with the exonic segments of one and only one gene. For many downstream analyses tools, such as DESeq, DESeq2, and EdgeR, only read-pairs in this category are used.

Ambig Gene: The read-pair overlaps with the exons of more than one gene.

No Gene: The read-pair does not overlap with the exons of any annotated gene.

No Gene, Intronic: The read-pair does not overlap with the exons of any annotated gene, but appears in a region that is bridged by an annotated splice junction.

No Gene, 1kb from gene: The read-pair does not overlap with the exons of any annotated gene, but is within 1 kilobase from the nearest annotated gene.

No Gene, 10kb from gene: The read-pair does not overlap with the exons of any annotated gene, but is within 10 kilobases from the nearest annotated gene.

No Gene, middle of nowhere: The read-pair does not overlap with the exons of any annotated gene, and is more than 10 kilobases from the nearest annotated gene.

See Also

build.plotter

[Package QoRTs version 0.2.11 Index]