makePlot.genebody.coverage {QoRTs}	R Documentation

Plot Gene-Body coverage distribution

Description

Plots Gene-Body coverage distribution

Usage

  makePlot.genebody.coverage(plotter, plot.medians,
                       plot.means = TRUE,  
                       debugMode, singleEndMode, ...)
  makePlot.genebody.coverage.UMQuartile(plotter, plot.medians, 
                       plot.means = TRUE, 
                       debugMode, singleEndMode, ...)
  makePlot.genebody.coverage.lowExpress(plotter, plot.medians, 
                       plot.means = TRUE, 
                       debugMode, singleEndMode, ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

plot.medians
A logical value indicating whether or not to plot the median average for each bamfile at the bottom of the plot. Overrides plot.means.

plot.means
A logical value indicating whether or not to plot the mean average for each bamfile at the bottom of the plot.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

x-axis: Gene-body quantile. By default this is broken up into 40 quantiles containing 2.5

y-axis: Rate at which reads falls into the given quantile of the genes' bodies.

makePlot.genebody.coverage plots the gene body coverage across all genes. makePlot.genebody.coverage.UMQuartile plots the gene body coverage across the genes that fall in the upper-middle quartile of total expression for each sample-run (excluding genes with zero reads). makePlot.genebody.coverage.lowExpress plots the gene body coverage across the genes that fall in the lower two quartiles of total expression for each sample-run (excluding genes with zero reads).

See Also

build.plotter

[Package QoRTs version 0.3.3 Index]