makePlot.legend.box {QoRTs}	R Documentation

Plot legend

Description

Plots the universal legend for a given plotter object.

Usage

makePlot.legend.box(plotter, debugMode, singleEndMode, ...)
makePlot.legend.over(position, plotter, debugMode, singleEndMode, ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

position
For makePlot.legend.over, a character string indicating the location where you want the legend to appear. This is passed to legend, and can be any keyword allowed by xy.coords. For example: "top", "topleft","bottomright", etc.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

makePlot.legend.box creates a new plot (opening the next graphics frame), and writes a small description of the given plotter type along with plotting a color-coded legend (if applicable).

makePlot.legend.over adds a legend to the current plotting frame.

See Also

build.plotter

[Package QoRTs version 0.3.26 Index]