makePlot.mapping.rates {QoRTs}	R Documentation

Plot mapping rates

Description

Plots the rates at which reads map to the genome.

Usage

makePlot.mapping.rates(plotter, plot.mm = NULL, 
                   y.counts.in.millions = TRUE, 
                   debugMode, singleEndMode,
                   ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

plot.mm
Plot multi-mapping rates. By default, it autodetects whether multimapping data was included in the original decoder.

y.counts.in.millions
Label/scale the y-axis in millions of reads.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

For each sample-run, this function plots the number of mapped reads and the rate at which reads map (if the total reads is provided).

See Also

build.plotter read.qc.results.data

[Package QoRTs version 0.3.9 Index]