makePlot.qual.pair {QoRTs}	R Documentation

Plot quality score by read cycle

Description

Plots the Phred quality score as a function of the read cycle for both reads.

Usage

  makePlot.qual.pair(plotter, y.name, r2.buffer = NULL, 
                 debugMode, singleEndMode, ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

y.name
The name of the quality score metric to plot. Must be one of:

"min"

"lowerQuartile"

"median"

"upperQuartile"

"max"

r2.buffer
Buffer space to place between the plotting of read 1 and read 2. By default this will choose a reasonable value.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

These plots display information about the phred quality score (y-axis) as a function of the position in the read (x-axis). Five statistics can be plotted: minimum, maximum, upper and lower quartiles, and median. These statistics are calculated individually for each bam file and each read position (ie, each plotted line corresponds to a bam file).

Note that the Phred score is always an integer, and as such these plots would normally be very difficult to read because lines would be plotted directly on top of one another. To reduce this problem, the lines are vertically offset from one another. Most plotters offset each line by lane.ID.

See Also

build.plotter

[Package QoRTs version 0.2.7 Index]