makePlot.splice.junction.loci.counts {QoRTs}	R Documentation

Splice Junction Loci Count Plot

Description

Plots the rate at which splice junction loci fall into various categories.

Usage

makePlot.splice.junction.loci.counts(plotter, 
                                 calc.rate = FALSE, 
                                 high.low.cutoff = 4,
                                 debugMode, singleEndMode, ...)

Arguments

plotter
A QoRT_Plotter reference object. See build.plotter.

calc.rate
Logical. If TRUE, the proportion of loci in each category will be calculated and plotted, rather than the raw number.

high.low.cutoff
(For advanced users only!) The cutoff between "high" and "low" expression junction loci. Note that this must match the cutoff used by the jarfile QC execution.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

singleEndMode
Logical. Determines whether the dataset consists of single-ended reads. By default this is determined from the data. Thus, this parameter should never need to be set by the user.

...
Arguments to be passed to methods, such as graphical parameters (see par).

Details

This function plots the number (y-axis) of splice junction loci of each type that appear in the bam-file's reads. Splice junctions are split into 4 groups, first by whether the splice junction appears in the transcript annotation gtf ("known" vs "novel"), and then by whether the splice junction has 4 or more reads covering it, or 1-3 reads ("Hi" vs "Lo").

See Also

build.plotter, makePlot.splice.junction.event

[Package QoRTs version 0.3.5 Index]