get.summary.table {QoRTs}	R Documentation

Get summary data tables

Description

Retrieves and compiles a summary data table.

Usage

  get.summary.table(res, outfile, debugMode);
  
  get.size.factors(res, 
       sf.method = c("DESeq2","DESeq2_GEO","TC",
                     "edgeR","edgeR_TMM","edgeR_UQ","edgeR_RLE"), 
       outfile, debugMode)

Arguments

res
A QoRT_QC_Results object, created by read.qc.results.data.

outfile
Optional. A file name where the table should be written.

sf.method
The size factor method to use. Note that most of these methods (except "TC") are reliant on external packages not included with QoRTs. You will need to install DESeq2 and edgeR to use these methods.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

Details

Returns summary data in table form.

See Also

read.qc.results.data

[Package QoRTs version 1.0.7 Index]