read.qc.results.data {QoRTs}	R Documentation

Reading QC results data

Description

Creates a QoRT_QC_Results object using a set of QC result data files.

Usage

  read.qc.results.data(infile.dir, 
                       decoder, 
                       decoder.files, 
                       calc.DESeq2 = FALSE, 
                       calc.edgeR = FALSE, 
                       debugMode)
                       
  completeAndCheckDecoder(decoder, decoder.files)

Arguments

infile.dir
REQUIRED. The base file directory where all the QC results data is stored.

decoder
A character vector or data.frame containing the decoder information. See details below.

decoder.files
Character vector. Either one or two character strings. Either decoder.files OR decoder must be set, never both. See details below.

calc.DESeq2
Logical. If TRUE, this function will attempt to load the DESeq2 package. If the DESeq2 package is found, it will then calculate DESeq2's geometric normalization factors (also known as "size factors") for each replicate and for each sample.

calc.edgeR
Logical. If TRUE, this function will attempt to load the edgeR package. If the edgeR package is found, it will then calculate all of edgeR's normalization factors (also known as "size factors") for each replicate and for each sample.

debugMode
Logical. If TRUE, debugging data will be printed to the console.

Details

read.qc.results.data reads in a full QoRTs dataset of multiple QoRTs QC runs and compiles them into a QoRTs_Results object.

completeAndCheckDecoder simply reads a decoder and "fills in" all missing parameters, returning a data.frame.

The "decoder" is used to describe each replicate/sample. The standard decoder is a data frame that has one row per replicate, with the following columns:

unique.ID: The base identifier for the individual replicate. Must be unique. lanebam.ID is a synonym.

lane.ID: (OPTIONAL) The identifier for the lane, run, or batch. The default is "UNKNOWN".

group.ID: (OPTIONAL) The identifier for the biological condition for the given replicate. The default is "UNKNOWN".

sample.ID: (OPTIONAL) The identifier for the specific biological replicate from which the replicate belongs. Note that this is distinct from "lanebam.ID" because in many RNA-Seq studies each "sample" can have multiple technical replicates, as multiple sequencing runs may be needed to acquire sufficient reads for analysis. By default, it is assumed that each replicate comes from a different sample, and sample.ID is set to equal unique.ID.

qc.data.dir: (OPTIONAL) This column indicates the subdirectory in which the replicate's QC data was written. If this column is missing, it is assumed to be equal to the unique.ID.

input.read.pair.count: (OPTIONAL) This column contains the number of read-pairs (or just reads, for single-end data), before alignment. This is used later to calculate mapping rate.

multi.mapped.read.pair.count: (OPTIONAL) This column contains the number of read-pairs (or just reads, for single-end data) that were multi-mapped. This must be included for multi-mapping rate to be calculated.

All the parameters except for unique.ID are optional. The decoder can even be supplied as a simple character vector, which is assumed to be the unique.ID's. All the other variables will be set to their default values.

Alternatively, the decoder can be supplied as a file given by the decoder.files parameter.

Dual Decoder: Optionally, two decoders can be supplied. In this case the first decoder should be the technical-replicate decoder and the second should be the biological-replicate decoder. The technical-replicate decoder should have one row per unique.ID, with the following columns:

unique.ID: The base identifier for the individual replicate. Must be unique!

lane.ID: (OPTIONAL) The identifier for the lane, run, or batch.

sample.ID: (OPTIONAL) The identifier for the specific biological replicate from which the replicate belongs. Note that this is distinct from "lanebam.ID" because in many RNA-Seq studies each "sample" can have multiple technical replicates, as multiple sequencing runs may be needed to acquire sufficient reads for analysis.

qc.data.dir: (OPTIONAL) This column indicates the subdirectory in which the replicate's QC data was written. If this column is missing, it is assumed to be equal to the lanebam.ID. Must be unique!

input.read.pair.count: (OPTIONAL) This column contains the number of input reads, before alignment. This is used later to calculate mapping rate.

multi.mapped.read.pair.count: (OPTIONAL) This column contains the number of reads that were multi-mapped. This must be included for multi-mapping rate to be calculated.

The biological-replicate decoder should have one row per sample.ID, with the following columns:

sample.ID: The identifier for the specific biological replicate from which the replicate belongs. Note that this is distinct from "unique.ID" because in many RNA-Seq studies each "sample" can have multiple technical replicates, as multiple sequencing runs may be needed to acquire sufficient reads for analysis.

group.ID (OPTIONAL): The identifier for the biological condition for the given replicate.

All decoders are allowed to contain other columns in addition to the ones listed here, so long as their names are distinct. Columns do not need to appear in any particular order, so long as they are named according to the specifications above.

See Also

build.plotter

Examples

  #Load the decoder from the example dataset:
  directory <- paste0(system.file("extdata/", 
                      package="QoRTsExampleData", 
                      mustWork=TRUE),"/");
  decoder.file <- system.file("extdata/decoder.txt", 
                              package="QoRTsExampleData", 
                              mustWork=TRUE);
  decoder.data <- read.table(decoder.file,
                             header=TRUE,
                             stringsAsFactors=FALSE);
  print(decoder.data);

##    sample.ID lane.ID unique.ID  qc.data.dir group.ID input.read.pair.count
## 1      SAMP1      L1 SAMP1_RG1 ex/SAMP1_RG1     CASE                465298
## 2      SAMP1      L2 SAMP1_RG2 ex/SAMP1_RG2     CASE                472241
## 3      SAMP1      L3 SAMP1_RG3 ex/SAMP1_RG3     CASE                500691
## 4      SAMP2      L1 SAMP2_RG1 ex/SAMP2_RG1     CASE                461405
## 5      SAMP2      L2 SAMP2_RG2 ex/SAMP2_RG2     CASE                467713
## 6      SAMP2      L3 SAMP2_RG3 ex/SAMP2_RG3     CASE                492322
## 7      SAMP3      L1 SAMP3_RG1 ex/SAMP3_RG1     CASE                485397
## 8      SAMP3      L2 SAMP3_RG2 ex/SAMP3_RG2     CASE                489859
## 9      SAMP3      L3 SAMP3_RG3 ex/SAMP3_RG3     CASE                516906
## 10     SAMP4      L1 SAMP4_RG1 ex/SAMP4_RG1     CTRL                460968
## 11     SAMP4      L2 SAMP4_RG2 ex/SAMP4_RG2     CTRL                468391
## 12     SAMP4      L3 SAMP4_RG3 ex/SAMP4_RG3     CTRL                484530
## 13     SAMP5      L1 SAMP5_RG1 ex/SAMP5_RG1     CTRL                469884
## 14     SAMP5      L2 SAMP5_RG2 ex/SAMP5_RG2     CTRL                475001
## 15     SAMP5      L3 SAMP5_RG3 ex/SAMP5_RG3     CTRL                494213
## 16     SAMP6      L1 SAMP6_RG1 ex/SAMP6_RG1     CTRL                452429
## 17     SAMP6      L2 SAMP6_RG2 ex/SAMP6_RG2     CTRL                458810
## 18     SAMP6      L3 SAMP6_RG3 ex/SAMP6_RG3     CTRL                477751

  #This command produces the example dataset used in all the other
  #  examples:
  res <- read.qc.results.data(directory, 
                              decoder = decoder.data,
                              calc.DESeq2 = TRUE, 
                              calc.edgeR = TRUE);

## Note: no multi.mapped.read.pair.count column found. This column is optional, but without it (depending on how your aligner implements multi-mapping) multi-mapping rates might not be plotted.

## reading QC.summary.txt files...

## done.

## [time: 2016-06-06 15:07:04],[elapsed: 0.03 secs]

## Autodetected Paired-End mode.

## reading QC.gc.byPair.txt.gz files...

## done.

## [time: 2016-06-06 15:07:04],[elapsed: 0.03 secs]

## reading QC.gc.byRead.txt.gz files...

## done.

## [time: 2016-06-06 15:07:04],[elapsed: 0.06 secs]

## reading QC.gc.byRead.vsBaseCt.txt.gz files...

## done.

## [time: 2016-06-06 15:07:04],[elapsed: 0.24 secs]

## reading QC.quals.r1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:04],[elapsed: 0.07 secs]

## reading QC.quals.r2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:04],[elapsed: 0.04 secs]

## reading QC.cigarOpDistribution.byReadCycle.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:05],[elapsed: 0.1 secs]

## reading QC.cigarOpDistribution.byReadCycle.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:05],[elapsed: 0.09 secs]

## reading QC.cigarOpLengths.byOp.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:05],[elapsed: 0.24 secs]

## reading QC.cigarOpLengths.byOp.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:05],[elapsed: 0.22 secs]

## reading QC.geneBodyCoverage.by.expression.level.txt.gz files...

## done.

## [time: 2016-06-06 15:07:05],[elapsed: 0.05 secs]

## reading QC.geneCounts.txt.gz files...

## done.

## [time: 2016-06-06 15:07:05],[elapsed: 0.29 secs]

## reading QC.insert.size.txt.gz files...

## done.

## [time: 2016-06-06 15:07:05],[elapsed: 0.06 secs]

## reading QC.NVC.raw.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:06],[elapsed: 0.1 secs]

## reading QC.NVC.raw.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:06],[elapsed: 0.07 secs]

## reading QC.NVC.lead.clip.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:07],[elapsed: 0.87 secs]

## reading QC.NVC.lead.clip.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:07],[elapsed: 0.85 secs]

## reading QC.NVC.tail.clip.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:08],[elapsed: 0.7 secs]

## reading QC.NVC.tail.clip.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:09],[elapsed: 0.74 secs]

## reading QC.NVC.minus.clipping.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:09],[elapsed: 0.07 secs]

## reading QC.NVC.minus.clipping.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:09],[elapsed: 0.09 secs]

## reading QC.chromCount.txt.gz files...

## done.

## [time: 2016-06-06 15:07:09],[elapsed: 0.03 secs]

## reading QC.biotypeCounts.txt.gz files...

## done.

## [time: 2016-06-06 15:07:09],[elapsed: 0.05 secs]

## reading QC.geneBodyCoverage.byExpr.avgPct.txt.gz files...

## done.

## [time: 2016-06-06 15:07:09],[elapsed: 0.06 secs]

## calculating secondary data:

## Calculating DESeq2 Normalization Factors (Geometric normalization)...

## Calculating edgeR Normalization Factors (all edgeR normalizations)...

## Calculating DESeq2 Normalization Factors (Geometric normalization)...

## Calculating edgeR Normalization Factors (all edgeR normalizations)...

## debug 3B

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 9 secs]

  #Note that DESeq2 and edgeR are required in order to 
  #  calculate the size factors using the options above.

  #You can also specify incomplete decoders, and use
  #  the following command to fill in the defaults:
  completeAndCheckDecoder(c("SAMP1","SAMP2",
                            "SAMP3","SAMP4",
                            "SAMP5","SAMP6"))

## Simple decoder found, list of sample/unique ID's. Building complete decoder.

## column 'qc.data.dir' not found in the decoder, assuming qc.data.dir = unique.ID

##   unique.ID sample.ID lane.ID group.ID qc.data.dir
## 1     SAMP1     SAMP1 UNKNOWN  UNKNOWN       SAMP1
## 2     SAMP2     SAMP2 UNKNOWN  UNKNOWN       SAMP2
## 3     SAMP3     SAMP3 UNKNOWN  UNKNOWN       SAMP3
## 4     SAMP4     SAMP4 UNKNOWN  UNKNOWN       SAMP4
## 5     SAMP5     SAMP5 UNKNOWN  UNKNOWN       SAMP5
## 6     SAMP6     SAMP6 UNKNOWN  UNKNOWN       SAMP6

  #You don't actually have to use completeAndCheckDecoder,
  #You can just pass the incomplete decoder directly to QoRTs.
  #For example, to load a small subset of the example data 
  #(without phenotype data):
  res <- read.qc.results.data(paste0(directory,"/ex/"), 
                              decoder = c("SAMP1_RG1","SAMP2_RG1",
                                          "SAMP3_RG1","SAMP4_RG1"));

## Simple decoder found, list of sample/unique ID's. Building complete decoder.
## column 'qc.data.dir' not found in the decoder, assuming qc.data.dir = unique.ID

## Note: no input.read.pair.count column found. This column is optional, but without it mapping rates cannot be calculated.

## Note: no multi.mapped.read.pair.count column found. This column is optional, but without it (depending on how your aligner implements multi-mapping) multi-mapping rates might not be plotted.

## reading QC.summary.txt files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.01 secs]

## Autodetected Paired-End mode.

## reading QC.gc.byPair.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.01 secs]

## reading QC.gc.byRead.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.01 secs]

## reading QC.gc.byRead.vsBaseCt.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.05 secs]

## reading QC.quals.r1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.02 secs]

## reading QC.quals.r2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.01 secs]

## reading QC.cigarOpDistribution.byReadCycle.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.02 secs]

## reading QC.cigarOpDistribution.byReadCycle.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.02 secs]

## reading QC.cigarOpLengths.byOp.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.05 secs]

## reading QC.cigarOpLengths.byOp.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.05 secs]

## reading QC.geneBodyCoverage.by.expression.level.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.01 secs]

## reading QC.geneCounts.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.07 secs]

## reading QC.insert.size.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.01 secs]

## reading QC.NVC.raw.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.03 secs]

## reading QC.NVC.raw.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.01 secs]

## reading QC.NVC.lead.clip.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:19],[elapsed: 0.2 secs]

## reading QC.NVC.lead.clip.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.2 secs]

## reading QC.NVC.tail.clip.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.15 secs]

## reading QC.NVC.tail.clip.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.17 secs]

## reading QC.NVC.minus.clipping.R1.txt.gz files...

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.02 secs]

## reading QC.NVC.minus.clipping.R2.txt.gz files...

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.03 secs]

## reading QC.chromCount.txt.gz files...

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.01 secs]

## reading QC.biotypeCounts.txt.gz files...

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.01 secs]

## reading QC.geneBodyCoverage.byExpr.avgPct.txt.gz files...

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.02 secs]

## calculating secondary data:

## debug 3B

## done.

## [time: 2016-06-06 15:07:20],[elapsed: 0.04 secs]

[Package QoRTs version 1.1.1 Index]